refactor: move some common patterns to separate files
